Refactor reduced dimensions
Reducing dimensions in indexing (i.e. array[:, 0] would reduce the
dimension of axis 1) is a bit tricky for block arrays, since you still
want to have the dimension there because the total number of dimensions
should match the number of dimensions in the blocks. For example, for a
block tensor with shape (1, 2, 3), the subtensors also have 3
dimensions. Indexing the block tensor with [:, :, 0] would normally reduce the
shape to (1, 2) but then the block tensor no longer has the same
dimension as the subtensor.

To fix this, -1 has been used to represent reduced dimensions and empty
tuples represent the corresponding empty block label.