Rename confusing blockarray.py to array.py
The array in blockarray.py doesn't have any blocks defined in it.
Instead the BlockArray object was a labelled array container for
arbitrary object and is used to contain the subtensors in BlockTensor.
As a result, it's been renamed to array.py and BlockArray has been
renamed to LabelledArray.py.