wip: Refactor the `.bshape` attribute to better handle collapsed dims
Previously collapsed dims would show up as an empty tuple in `.bshape`.
This has been changed to be a single int representing the size of that
dimension.

This might be a good strategy for `.bshape` to represent arbitrarily
nested arrays in a recursive fasion. if an axis size is a tuple of ints,
then that axis is split into multiple arrays. If an axis size is a
single int then that axis has a single dimension. I haven't thought out
how multiple levels of nesting would work for arrays in details so this
might only be good for the single nesting level considered here.